Switching the register components to page routes; Step 1 and 2 changed to adapt new system on back
