+++ locales/en-US/register.json
@@ -14,7 +14,8 @@
         "fornameAlert": "Insert only you last name, no spaces",
         "emailAlert": "Insert a valid e-mail!",
         "maxYearExceeded": "Age must be 16 or older.",
-        "nextButtonText": "Next"
+        "nextButtonText": "Next",
+        "emailSendedMessage": "We sended a confirmation link to your e-mail. Access it to proceed with the registration."
     },
 
     "step2": {
@@ -24,6 +25,9 @@
         "usernameAlert": "Use only alphanumeric characters and some special ones ('.', '_' e '-').",
         "passwordAlert": "Password must contain at least 8 characters and no spaces.",
         "confirmPasswordAlert": "Passwords must match!",
-        "registerButtonText": "Register"
+        "registerButtonText": "Register",
+        "invalidGuid": "Session expired, please try again.",
+        "userAlreadyRegisteredErrorMessage": "This link has expired. Access your created account on the home page.",
+        "backButtonText": "Back to the home page"
     }
 }
