+++ services/api.js
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+export const api = axios.create({
+    baseURL: ''
+});
